🔨 Vehicle refactor 🔨
